requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
asyncpg==0.30.0
httpx==0.28.1
python-dotenv==1.1.1
